backend/pyproject.toml
[project]
name = "parking-backend"
version = "0.1.0"
description = "Reducing parking congestion using real-time space availability and digital reservation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.3",
    "django-allauth>=65.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-money>=3.5.4",
    "django-restql>=0.16.2",
    "django-types>=0.21.0",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "pytest-django>=4.11.1",
    "requests>=2.32.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
